requirements.txt
RPi.GPIO>=0.7.1
adafruit-circuitpython-servokit>=1.3.0
Pillow>=9.0.0
spidev>=3.5
board
busio
